Remove old Metric.Helper and renew the new one to Metric.Helper
